+++ lib/binaries.sh
@@ -72,14 +72,7 @@ install_nodejs() {
   local code resolve_result
 
   if [[ -z "$version" ]]; then
-    # Node.js 18+ is incompatible with ubuntu:18 (and thus heroku-18) because of a libc mismatch:
-    # node: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.28' not found (required by node)
-    # Fallback to a 16.x default for heroku-18 until heroku-18 or Node.js 16.x are EOL.
-    if [[ "$STACK" == "heroku-18" ]]; then
-      version="16.x"
-    else
       version="18.x"
-    fi
   fi
 
   if [[ -n "$NODE_BINARY_URL" ]]; then